requirements.txt
google-generativeai
python-dotenv
cryptography
icalendar
pytz 